chore: refactor and remove "should" from test description
